feat: 애플리케이션 이벤트 추가

+++ src/main/java/com/jolupbisang/demo/infrastructure/meeting/client/dto/response/DiarizedResponse.java
@@ -21,8 +21,8 @@ public record DiarizedResponse(
     }
 
     public record Word(
-            float start,
-            float end,
+            int start,
+            int end,
             String text,
             String lang
     ) {
